Extract transfer_records to class
 * Ensures not globally available, only where needed
   (One place in QT)
 * Ensures not private, so can be called outside model class
