+++ lib/rails_core_extensions.rb
@@ -6,6 +6,7 @@ module RailsCoreExtensions
   require 'rails_core_extensions/breadcrumb'
   require 'rails_core_extensions/position_initializer'
   require 'rails_core_extensions/time_with_zone'
+  require 'rails_core_extensions/transfer_records'
   require 'rails_core_extensions/active_support_concern'
   require 'rails_core_extensions/concurrency'
